chore(Liveness): UX Changes (#5105)
* fix(liveness): add custom signer to client (#5063)

* chore(liveness): remove one second oval match criteria (#5093)

* chore(liveness): update face match algorithm (#5085)

* chore(liveness): Add one second pause between match and flashing lights  (#5102)
---------
