+++ .changeset/happy-crabs-give.md
@@ -1,0 +1,5 @@
+---
+"@aws-amplify/ui-react-liveness": patch
+---
+
+chore(liveness): Add a one second pause between face match and flashing lights, giving time to read "Hold still" text.
