+++ packages/react-liveness/jest.config.ts
@@ -11,10 +11,10 @@ const config: Config = {
   ],
   coverageThreshold: {
     global: {
-      branches: 82,
+      branches: 80,
       functions: 82,
-      lines: 91,
-      statements: 91,
+      lines: 90,
+      statements: 90,
     },
   },
   testPathIgnorePatterns: [],
