+++ packages/react-liveness/src/components/FaceLivenessDetector/service/types/liveness.ts
@@ -122,7 +122,6 @@ export enum IlluminationState {
 export enum FaceMatchState {
   MATCHED = 'MATCHED',
   TOO_FAR = 'TOO FAR',
-  TOO_CLOSE = 'TOO CLOSE',
   CANT_IDENTIFY = 'CANNOT IDENTIFY',
   FACE_IDENTIFIED = 'ONE FACE IDENTIFIED',
   TOO_MANY = 'TOO MANY FACES',
