+++ packages/react-liveness/src/components/FaceLivenessDetector/service/utils/__tests__/liveness.test.ts
@@ -3,29 +3,96 @@ import {
   drawLivenessOvalInCanvas,
   estimateIllumination,
   fillOverlayCanvasFractional,
+  generateBboxFromLandmarks,
   getColorsSequencesFromSessionInformation,
   getFaceMatchState,
-  getFaceMatchStateInLivenessOval,
   getOvalDetailsFromSessionInformation,
   isCameraDeviceVirtual,
   isFaceDistanceBelowThreshold,
 } from '../liveness';
 import {
   getMockContext,
+  mockCameraDevice,
   mockFace,
   mockOvalDetails,
   mockSessionInformation,
-  mockCameraDevice,
 } from '../__mocks__/testUtils';
 import {
   Face,
   FaceMatchState,
   IlluminationState,
   LivenessErrorState,
-  LivenessOvalDetails,
 } from '../../types';
 
 const context = getMockContext();
+
+const MOCK_TIMESTAMP = 1640995200000;
+
+// Values generated by passing mockFace into generateBboxFromLandmarks and printing intermediate values
+const mockFaceHeight = 0; // 1.68 * [(ALPHA * pd + GAMMA * fh) / 2];
+const mockFaceCenterY = 200; // eyeCenter[1];
+
+const mockAboveFrameFace: Face = {
+  height: 215,
+  left: 200,
+  leftEar: [420, 155],
+  leftEye: [360, 110],
+  mouth: [310, 200],
+  nose: [310, 140],
+  probability: 90,
+  rightEar: [215, 170],
+  rightEye: [260, 115],
+  timestampMs: MOCK_TIMESTAMP,
+  top: 70,
+  width: 215,
+};
+
+const mockCloselyMatchedFace: Face = {
+  top: 46.751117706298814,
+  left: -5.630989074707031,
+  width: 572.6491546630859,
+  height: 572.7001762390136,
+  timestampMs: MOCK_TIMESTAMP,
+  probability: 90,
+  rightEye: [187.4885559082032, 218.63224029541013],
+  leftEye: [433.75030517578125, 227.35191345214847],
+  mouth: [313.39622497558594, 485.65040588378895],
+  nose: [330.15281677246094, 371.40689849853516],
+  rightEar: [28.786602020263672, 237.78337478637698],
+  leftEar: [515.0150299072266, 262.62935638427734],
+};
+
+const mockMatchedFace: Face = {
+  height: 317.5821685791015,
+  left: 165.49339294433594,
+  leftEar: [445, 264.4227600097656],
+  leftEye: [384.70985412597656, 247.56172180175784],
+  mouth: [327.9991149902344, 393.5927200317383],
+  nose: [325.2409362792969, 333.89575958251953],
+  probability: 90,
+  rightEar: [199.34513092041016, 272.49412536621094],
+  rightEye: [261.6714286804199, 250.6684875488281],
+  timestampMs: MOCK_TIMESTAMP,
+  top: 155.82656860351565,
+  width: 317.60498046875,
+};
+const mockMatchedFaceCenterX = 323.19064140319824;
+const mockMatchedFaceOcularWidth = 253.17949214956178;
+
+const mockTurnedFace: Face = {
+  height: 215.05796432495114,
+  left: 235.08899688720703,
+  leftEar: [433.8389587402344, 233.52020263671872],
+  leftEye: [379.2174530029297, 205.73320388793942],
+  mouth: [340.50960540771484, 304.9227523803711],
+  nose: [337.0632553100586, 257.2260284423828],
+  probability: 90,
+  rightEar: [250.12165069580078, 249.12113189697263],
+  rightEye: [292.9753875732422, 215.29994964599607],
+  timestampMs: MOCK_TIMESTAMP,
+  top: 158.50336074829102,
+  width: 215.0579833984375,
+};
 
 describe('Liveness Helper', () => {
   describe('getOvalDetailsFromSessionInformation', () => {
@@ -39,134 +106,6 @@ describe('Liveness Helper', () => {
       expect(ovalParameters.centerY).toBe(4);
       expect(ovalParameters.width).toBe(1);
       expect(ovalParameters.height).toBe(2);
-    });
-  });
-
-  describe('getFaceMatchStateInLivenessOval', () => {
-    it('should parse sessionInformation and return oval parameter attributes', () => {
-      const face: Face = mockFace;
-      const ovalDetails: LivenessOvalDetails = mockOvalDetails;
-      const initialIntersection: number = 0.3;
-      const sessionInfo = mockSessionInformation;
-
-      const { faceMatchState, faceMatchPercentage } =
-        getFaceMatchStateInLivenessOval(
-          face,
-          ovalDetails,
-          initialIntersection,
-          sessionInfo
-        );
-
-      expect(faceMatchState).toBe(FaceMatchState.OFF_CENTER);
-      expect(faceMatchPercentage).toBe(0);
-    });
-
-    it('should return TOO CLOSE', () => {
-      const face: Face = {
-        height: 586,
-        width: 586,
-        left: -6,
-        top: 21,
-        timestampMs: 1,
-        rightEye: [185, 231],
-        leftEye: [427, 224],
-        mouth: [330, 493],
-        nose: [334, 400],
-        rightEar: [36, 243],
-        leftEar: [518, 217],
-      };
-      const ovalDetails: LivenessOvalDetails = {
-        height: 465,
-        width: 288,
-        flippedCenterX: 320,
-        centerX: 320,
-        centerY: 240,
-      };
-      const initialIntersection: number = 0.3;
-      const sessionInfo = mockSessionInformation;
-
-      const { faceMatchState } = getFaceMatchStateInLivenessOval(
-        face,
-        ovalDetails,
-        initialIntersection,
-        sessionInfo
-      );
-
-      expect(faceMatchState).toBe(FaceMatchState.TOO_CLOSE);
-    });
-
-    // Note: if this test ever fails due to face detection updates just add a console log for face and do a check and copy the value here
-    it('should return matched', () => {
-      const face: Face = {
-        top: 89.82275009155273,
-        left: 188.85473251342773,
-        width: 375.7287788391113,
-        height: 375.66087722778315,
-        timestampMs: 1696575105415,
-        rightEye: [291.9423294067383, 241.10103607177734],
-        leftEye: [435.9743881225586, 264.43485260009766],
-        mouth: [336.0434341430664, 417.43167877197266],
-        nose: [336.7517852783203, 387.286376953125],
-        rightEar: [236.0750961303711, 193.77853393554685],
-        leftEar: [520.6099700927734, 238.4494781494141],
-      };
-      const ovalDetails: LivenessOvalDetails = {
-        centerX: 320,
-        centerY: 240,
-        flippedCenterX: 320,
-        height: 512,
-        width: 316,
-      };
-      const initialIntersection: number = 0.3;
-      const sessionInfo = mockSessionInformation;
-
-      const { faceMatchState, faceMatchPercentage } =
-        getFaceMatchStateInLivenessOval(
-          face,
-          ovalDetails,
-          initialIntersection,
-          sessionInfo
-        );
-
-      expect(faceMatchState).toBe(FaceMatchState.MATCHED);
-      expect(faceMatchPercentage).toBe(100);
-    });
-
-    it('should return TOO FAR', () => {
-      const face: Face = {
-        top: 229.4625244140625,
-        left: 275.884765625,
-        width: 135.34274291992188,
-        height: 101.50827026367188,
-        timestampMs: 1683674844363,
-        probability: 0.9974300265312195,
-        rightEye: [372.5564064979553, 258.19776356220245],
-        leftEye: [318.5161700248718, 253.94269466400146],
-        mouth: [339.64158596098423, 298.5959941148758],
-        nose: [342.7122294306755, 277.0021167397499],
-        rightEar: [400.5564064979553, 258.19776356220245],
-        leftEar: [300.5161700248718, 253.94269466400146],
-      };
-      const ovalDetails: LivenessOvalDetails = {
-        flippedCenterX: 346,
-        centerX: 294,
-        centerY: 215,
-        width: 316,
-        height: 512,
-      };
-      const initialIntersection: number = 0.3;
-      const sessionInfo = mockSessionInformation;
-
-      const { faceMatchState, faceMatchPercentage } =
-        getFaceMatchStateInLivenessOval(
-          face,
-          ovalDetails,
-          initialIntersection,
-          sessionInfo
-        );
-
-      expect(faceMatchState).toBe(FaceMatchState.TOO_FAR);
-      expect(faceMatchPercentage).toBe(0);
     });
   });
 
@@ -208,6 +147,83 @@ describe('Liveness Helper', () => {
       );
 
       expect(faceMatchState).toBe(FaceMatchState.TOO_MANY);
+    });
+  });
+
+  describe('generateBboxfromLandmarks', () => {
+    it(`should return face box bottom as 'bottom' when face is within frame`, () => {
+      const frameHeight = 480;
+      const { bottom } = generateBboxFromLandmarks(
+        mockFace,
+        mockOvalDetails,
+        frameHeight
+      );
+
+      const faceBoxBottom = mockFaceCenterY + mockFaceHeight / 2;
+      expect(bottom).toEqual(faceBoxBottom); // expect calculated value to be returned as it is within frame
+    });
+
+    it(`should return 'frameHeight' as 'bottom' when bottom of face is below frame`, () => {
+      const frameHeight = 100;
+      const { bottom } = generateBboxFromLandmarks(
+        mockFace,
+        mockOvalDetails,
+        frameHeight
+      );
+      expect(bottom).toEqual(frameHeight); // expect frameHeight to be returned as it is smaller than the calculated value
+    });
+
+    it(`should return face box top as 'top' when face is within frame`, () => {
+      const frameHeight = 480;
+      const { top } = generateBboxFromLandmarks(
+        mockFace,
+        mockOvalDetails,
+        frameHeight
+      );
+
+      const faceBoxTop = mockFaceCenterY - mockFaceHeight / 2;
+
+      expect(top).toEqual(faceBoxTop);
+    });
+
+    it(`should return 0 as 'top' when top of face is above frame`, () => {
+      const frameHeight = 480;
+      const { top } = generateBboxFromLandmarks(
+        mockAboveFrameFace,
+        mockOvalDetails,
+        frameHeight
+      );
+
+      expect(top).toEqual(0);
+    });
+
+    it(`should return correct 'left' and 'right' values when face is matched`, () => {
+      const frameHeight = 480;
+
+      const { left, right } = generateBboxFromLandmarks(
+        mockMatchedFace,
+        mockOvalDetails,
+        frameHeight
+      );
+
+      const faceBoxRight =
+        mockMatchedFaceCenterX + mockMatchedFaceOcularWidth / 2;
+      const faceBoxLeft =
+        mockMatchedFaceCenterX - mockMatchedFaceOcularWidth / 2;
+      expect(left).toEqual(faceBoxLeft); // expect left = cx - ow / 2
+      expect(right).toEqual(faceBoxRight); // expect right = cx + ow / 2
+    });
+
+    it(`should return correct 'left' and 'right' values when face is turned`, () => {
+      const frameHeight = 480;
+      const { left, right } = generateBboxFromLandmarks(
+        mockTurnedFace,
+        mockOvalDetails,
+        frameHeight
+      );
+
+      expect(left).toEqual(mockTurnedFace.rightEar[0]); // expect right ear to be used to limit left edge of face
+      expect(right).toEqual(mockTurnedFace.leftEar[0]); // expect left ear to be used to limit right edge of face
     });
   });
 
@@ -266,20 +282,7 @@ describe('Liveness Helper', () => {
     });
 
     it('should return false and error if above threshold', async () => {
-      const mockCloseFace: Face = {
-        height: 100,
-        width: 100,
-        left: 150,
-        top: 200,
-        timestampMs: 1,
-        rightEye: [0, 100],
-        leftEye: [150, 100],
-        mouth: [100, 100],
-        nose: [100, 100],
-        rightEar: [0, 100],
-        leftEar: [150, 100],
-      };
-      mockBlazeFace.detectFaces.mockResolvedValue([mockCloseFace]);
+      mockBlazeFace.detectFaces.mockResolvedValue([mockCloselyMatchedFace]);
 
       const result = await isFaceDistanceBelowThreshold({
         faceDetector: mockBlazeFace,
