+++ packages/react-liveness/src/components/FaceLivenessDetector/service/utils/index.ts
@@ -2,5 +2,6 @@ export * from './blazefaceFaceDetection';
 export * from './videoRecorder';
 export * from './support';
 export * from './liveness';
+export * from './getFaceMatchStateInLivenessOval';
 export * from './streamProvider';
 export * from './freshnessColorDisplay';
